refactor: rename calendar -> schedule
