--- escahp/src/main/java/com/esca/escahp/calendar/repository/CalendarRepository.java
@@ -1,11 +1,0 @@
-package com.esca.escahp.calendar.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.esca.escahp.calendar.entity.Calendar;
-
-@Repository
-public interface CalendarRepository extends JpaRepository<Calendar, Long> {
-
-}
