+++ escahp/src/main/java/com/esca/escahp/schedule/I_ScheduleService.java
@@ -1,0 +1,4 @@
+package com.esca.escahp.schedule;
+
+public interface I_ScheduleService {
+}
