+++ escahp/src/main/java/com/esca/escahp/schedule/ScheduleService.java
@@ -1,0 +1,7 @@
+package com.esca.escahp.schedule;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ScheduleService {
+}
