+++ escahp/src/main/java/com/esca/escahp/schedule/repository/ScheduleRepository.java
@@ -1,0 +1,11 @@
+package com.esca.escahp.schedule.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.esca.escahp.schedule.entity.Schedule;
+
+@Repository
+public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
+
+}
