+++ escahp/src/main/java/com/esca/escahp/timetable/TimeTableService.java
@@ -1,17 +1,17 @@
 package com.esca.escahp.timetable;
 
-import com.esca.escahp.calendar.repository.CalendarRepository;
+import com.esca.escahp.schedule.repository.ScheduleRepository;
 import com.esca.escahp.timetable.repository.TimeTableRepositorty;
 
 import org.springframework.stereotype.Service;
 
 @Service
 public class TimeTableService implements I_TimeTableService {
-    private final CalendarRepository calendarRepository;
+    private final ScheduleRepository scheduleRepository;
     private final TimeTableRepositorty timeTableRepositorty;
 
-    public TimeTableService(CalendarRepository keeperRepository, TimeTableRepositorty timeTableRepositorty) {
-        this.calendarRepository = keeperRepository;
+    public TimeTableService(ScheduleRepository keeperRepository, TimeTableRepositorty timeTableRepositorty) {
+        this.scheduleRepository = keeperRepository;
         this.timeTableRepositorty = timeTableRepositorty;
     }
 }
